+++ torch/_dynamo/guards.py
@@ -82,6 +82,7 @@ from torch._guards import (
     Source,
     StorageOverlap,
 )
+from torch._guards.install import install_guard  # noqa: F401
 from torch._guards.manager import (  # noqa: F401
     DeletedGuardManagerWrapper,
     dunder_attrs_assumed_constants,
@@ -5139,32 +5140,6 @@ def make_dupe_guard(
             # However, this should always be a sound guard to add here.
             return functools.partial(GuardBuilder.DUPLICATE_INPUT, source_b=dupe_source)
     return None
-
-
-def install_guard(*guards: Guard, skip: int = 0) -> None:
-    """
-    Add dynamo guards to the current tracing context.
-
-    Args:
-        guards: guard(s) to add
-        skip: number of stack frames to ignore for debug stack trace
-    """
-    from torch._guards import TracingContext
-
-    guards_context = TracingContext.get().guards_context
-    if guards_context.skip_install:
-        return
-
-    collect_debug_stack = guards_log.isEnabledFor(
-        logging.DEBUG
-    ) or verbose_guards_log.isEnabledFor(logging.DEBUG)
-    add = guards_context.dynamo_guards.add
-    for guard in guards:
-        if not isinstance(guard, Guard):
-            raise AssertionError(f"Expected Guard, got {type(guard)}")
-        if is_from_skip_guard_source(guard.originating_source):
-            continue
-        add(guard, collect_debug_stack=collect_debug_stack, skip=skip + 1)
 
 
 # ---------------------------------------------------------------------------
